backend/pyproject.toml
[tool.poetry]
name = "fittracker-backend"
version = "0.1.0"
description = "Backend for FitTracker application"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.0"
pydantic = {extras = ["email"], version = "^2.5.3"}
tenacity = "^8.2.3" # For retry mechanisms, e.g. DB connection

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
httpx = "^0.25.2" # For testing FastAPI endpoints

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
